pyproject.toml
[project]
name = "restaurant-picker"
version = "0.1.0"
description = "A simple web app to find restaurants by country cuisine"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]
